Testing-requirements.txt
Flask-Testing
fixture
Rednose>=0.4.1
unittest2>=0.5.1
coveralls
mock>=1.0.1
coverage>=3.7.1
nose
blinker
